# swagger/swagger.json
swagger: "2.0"
info:
  description: "API Rest desarrollada en node.js - Modulo estudiantes"
  version: "1.0.0"
  title: "API REST ESTUDIANTES"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "godaycn87@outlook.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "estudiante-app.herokuapp.com"
schemes:
  - "http"
  - "https"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
basePath: "/"
...
paths:
...
  /traerMaterias/{id}:
    get:
      tags:
      - "Calificaciones"
      summary: "RPC de traerMaterias"
      description: ""
      operationId: "traerMaterias"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID de docente"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "300":
          description: "Usted no posee materias"
        "400":
          description: "El id no puede estar vacio"
        "500":
          description: "Id invalido"
      security:
        - Bearer: []
...
  /inscribirEstudianteCursada:
    post:
      tags:
      - "inscripciones"
      summary: "inscribirEstudianteCursada"
      description: ""
      operationId: "inscribirEstudianteCursada"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "notas"
        in: "body"
        description: "Parametros listado para inscribirEstudianteCursada"
        required: true
        schema:
          $ref: '#/definitions/notasArchivo'
      responses:
        "200":
          description: "OK"
        "400":
          description: "El body no puede estar vacio"
        "300":
          description: "No se encontraron alumnos"
        "401":
          description: "No se pudo completar la operación"
        "500":
          description: "datos invalido"
      security:
        - Bearer: []                            
  /traerMateriasParaInscripcion/{idEstudiante}:
    get:
      tags:
      - "inscripciones"
      summary: "traerMateriasParaInscripcion"
      description: ""
      operationId: "traerMateriasParaInscripcion"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "idEstudiante"
        in: "path"
        description: "ID de estudiante"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "OK"
        "400":
          description: "El idEstudiante no puede estar vacio"
        "500":
          description: "Falla al insertar"
      security:
        - Bearer: []    
  /bajaInscripcionMateria/{idalumnosCursada}:
    delete:
      tags:
      - "inscripciones"
      summary: "bajaInscripcionMateria"
      description: ""
      operationId: "bajaInscripcionMateria"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "idalumnosCursada"
        in: "path"
        description: "ID de estudiante"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "OK"
        "400":
          description: "El body no puede estar vacio"
        "406":
          description: "INSCRIPCION FUERA DE FECHA PARA ELIMINAR O NO EXISTE"
        "500":
          description: "datos invalido"
      security:
        - Bearer: []            
  /traerInscripcionesEstudianteCursada/{idEstudiante}:
    get:
      tags:
      - "inscripciones"
      summary: "traerInscripcionesEstudianteCursada"
      description: ""
      operationId: "traerInscripcionesEstudianteCursada"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "idEstudiante"
        in: "path"
        description: "ID de estudiante"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "OK"
        "400":
          description: "El idEstudiante no puede estar vacio"
        "500":
          description: "datos invalido"
      security:
        - Bearer: []    
...
definitions:
...
  notasArchivo:
    type: "object"
    properties:
      idDocente:
        type: "integer"
        format: "int64"
      idMateria:
        type: "integer"
        format: "int64"
      asistencia:
        type: "integer"
        format: "int64"
      examenesidExamenes:
        type: "integer"
        format: "int64"
      idInscriptosExamen:
        type: "integer"
        format: "int64"
      nota:
        type: "integer"
        format: "int64"
      recordatorio:
        type: "integer"
        format: "int64"
    xml:
      name: "notas"  
...